api/requirements.txt
Django
djangorestframework
ipython
pillow
psycopg2
reportlab
